examples/requirements.txt
numpy>=1.24.0
pandas>=1.5.0
Pillow>=9.0.0
jax>=0.4.13
jaxlib>=0.4.13 